測試credit shader

+++ Assets/HelloPico2/Data/VFX/GGDog/Other/VertexColor.shader
@@ -45,9 +45,9 @@ Shader "Unlit/VertexColor"
                 return o;
             }
 
-            fixed4 frag (v2f i) : SV_Target
+            float4 frag (v2f i) : SV_Target
             {
-                fixed4 col = tex2D(_MainTex, i.uv)*i.color;
+                float4 col = tex2D(_MainTex, i.uv)*i.color;
                 return col;
             }
             ENDCG
